requirements.txt
# requirements.txt (Updated for NDIS features)
# Web Framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# AI/ML
openai>=1.3.0
google-generativeai>=0.3.0

# Database
sqlalchemy>=2.0.0

# Document Processing
reportlab>=4.0.0
python-docx>=0.8.11
jinja2>=3.1.0
markdown>=3.5.0
PyMuPDF>=1.23.0  # For PDF text extraction

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.4.0
email-validator>=2.1.0

# HTTP and API
requests>=2.31.0
httpx>=0.25.0

# Development
python-dotenv>=1.0.0
pytest>=7.4.0
black>=23.9.0

docxtpl-0.20.1
docx2pdf-0.1.8

docxtpl>=0.16.7
pypdf2-3.0.1

google-generativeai-0.4.6
